# doc/api.yaml
openapi: 3.0.0
paths:
  /user:
    get:
      operationId: UsersController_getAll
      summary: Get all users
      description: Gets all users
      parameters: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
      tags:
        - Users
    post:
      operationId: UsersController_create
      summary: Create user
      description: Creates a new user
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: The user has been created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request. body does not contain required fields
      tags:
        - Users
  /user/{id}:
    get:
      operationId: UsersController_getById
      summary: Get single user by id
      description: Get single user by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request. id is invalid (not uuid)
        '404':
          description: User not found
      tags:
        - Users
    put:
      operationId: UsersController_update
      summary: Update a user's password
      description: Updates a user's password by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePasswordDto'
      responses:
        '200':
          description: The user has been updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request. userId is invalid (not uuid)
        '403':
          description: oldPassword is wrong
        '404':
          description: User not found
      tags:
        - Users
    delete:
      operationId: UsersController_delete
      summary: Delete user
      description: Delete user by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '204':
          description: The user has been deleted
        '400':
          description: Bad request. User id is invalid (not uuid)
        '404':
          description: User not found
      tags:
        - Users
  /track:
    get:
      operationId: TracksController_findAll
      summary: Get tracks list
      description: Gets all library tracks list
      parameters: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Track'
      tags:
        - Tracks
    post:
      operationId: TracksController_create
      summary: Add new track
      description: Add new track information
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTrackDto'
      responses:
        '201':
          description: The track has been created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        '400':
          description: Bad request. Body does not contain required fields
      tags:
        - Tracks
  /track/{id}:
    get:
      operationId: TracksController_getById
      summary: Get single track by id
      description: Get single track by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        '400':
          description: Bad request. Track id is invalid (not uuid)
        '404':
          description: Track not found
      tags:
        - Tracks
    put:
      operationId: TracksController_update
      summary: Update track information
      description: Update track information by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTrackDto'
      responses:
        '200':
          description: The track information has been updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        '400':
          description: Bad request. Track id is invalid (not uuid)
        '404':
          description: Track not found
      tags:
        - Tracks
    delete:
      operationId: TracksController_delete
      summary: Delete track
      description: Delete track from library by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '204':
          description: Track has been deleted
        '400':
          description: Bad request. Track id is invalid (not uuid)
        '404':
          description: Track not found
      tags:
        - Tracks
  /artist:
    get:
      operationId: ArtistsController_findAll
      summary: Get artists list
      description: Gets all artists list
      parameters: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Artist'
      tags:
        - Artists
    post:
      operationId: ArtistsController_create
      summary: Add new artist
      description: Add new artist information
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateArtistDto'
      responses:
        '201':
          description: The artist has been created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        '400':
          description: Bad request. Body does not contain required fields
      tags:
        - Artists
  /artist/{id}:
    get:
      operationId: ArtistsController_getById
      summary: Get single artist by id
      description: Get single artist by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        '400':
          description: Bad request. Artist id is invalid (not uuid)
        '404':
          description: Atist not found
      tags:
        - Artists
    put:
      operationId: ArtistsController_update
      summary: Update artist information
      description: Update artist information by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateArtistDto'
      responses:
        '200':
          description: The artist information has been updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        '400':
          description: Bad request. Artist id is invalid (not uuid)
        '404':
          description: Artist not found
      tags:
        - Artists
    delete:
      operationId: ArtistsController_delete
      summary: Delete artist
      description: Delete artist from library by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '204':
          description: Artist has been deleted
        '400':
          description: Bad request. Artist id is invalid (not uuid)
        '404':
          description: Artist not found
      tags:
        - Artists
  /album:
    get:
      operationId: AlbumsController_findAll
      summary: Get albums list
      description: Gets all library albums list
      parameters: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Album'
      tags:
        - Albums
    post:
      operationId: AlbumsController_create
      summary: Add new album
      description: Add new album information
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAlbumDto'
      responses:
        '201':
          description: The album has been created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        '400':
          description: Bad request. body does not contain required fields
      tags:
        - Albums
  /album/{id}:
    get:
      operationId: AlbumsController_getById
      summary: Get single album by id
      description: Get single album by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        '400':
          description: Bad request. Album id is invalid (not uuid)
        '404':
          description: Album not found
      tags:
        - Albums
    put:
      operationId: AlbumsController_update
      summary: Update album information
      description: Update album information by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAlbumDto'
      responses:
        '200':
          description: The album information has been updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        '400':
          description: Bad request. Album id is invalid (not uuid)
        '404':
          description: Album not found
      tags:
        - Albums
    delete:
      operationId: AlbumsController_delete
      summary: Delete album
      description: Delete album from library by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '204':
          description: Album has been deleted
        '400':
          description: Bad request. Album id is invalid (not uuid)
        '404':
          description: Album not found
      tags:
        - Albums
  /favs:
    get:
      operationId: FavoritesController_findAll
      summary: Get favorites
      description: Gets all favorites movies, tracks and books
      parameters: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseFavoritesDto'
      tags:
        - Favorites
  /favs/track/{id}:
    post:
      operationId: FavoritesController_addTrack
      summary: Add track to the favorites
      description: Add track to the favorites
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: The track has been created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        '400':
          description: Bad request. Track id is invalid (not uuid)
        '404':
          description: Track with id not found
      tags:
        - Favorites
    delete:
      operationId: FavoritesController_deleteTrack
      summary: Delete track
      description: Delete track from favorites by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '204':
          description: Track has been deleted from favorites
        '400':
          description: Bad request. Track id is invalid (not uuid)
        '404':
          description: Track not found
      tags:
        - Favorites
  /favs/artist/{id}:
    post:
      operationId: FavoritesController_addArtist
      summary: Add artist to the favorites
      description: Add artist to the favorites
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: The artist has been created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        '400':
          description: Bad request. Artist id is invalid (not uuid)
        '404':
          description: Artist with id not found
      tags:
        - Favorites
    delete:
      operationId: FavoritesController_deleteArtist
      summary: Delete artist from favorites
      description: Delete artist from favorites by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '204':
          description: Artist has been deleted from the favorites
        '400':
          description: Bad request. Artist id is invalid (not uuid)
        '404':
          description: Artist not found
      tags:
        - Favorites
  /favs/album/{id}:
    post:
      operationId: FavoritesController_addAlbum
      summary: Add album to the favorites
      description: Add album to the favorites
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: The album has been created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        '400':
          description: Bad request. Album id is invalid (not uuid)
        '404':
          description: Album with id not found
      tags:
        - Favorites
    delete:
      operationId: FavoritesController_deleteAlbum
      summary: Delete album from favorites
      description: Delete album from favorites by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '204':
          description: Album has been deleted from favorites
        '400':
          description: Bad request. Album id is invalid (not uuid)
        '404':
          description: Album not found
      tags:
        - Favorites
  /:
    get:
      operationId: AppController_getHello
      parameters: []
      responses:
        '200':
          description: ''
info:
  title: Home Library Service
  description: Home music library service
  version: '1.0'
  contact: {}
tags: []
servers:
  - url: /
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
          format: uuid
        login:
          type: string
          example: TestUser
        version:
          type: number
          example: 1
          description: Entity version. Auto-incremented on each update
        createdAt:
          type: number
          example: 1655000000
        updatedAt:
          type: number
          example: 1655000000
      required:
        - id
        - login
        - version
        - createdAt
        - updatedAt
    CreateUserDto:
      type: object
      properties:
        login:
          type: string
          example: TestUser
          format: uuid
        password:
          type: string
          example: test-password
      required:
        - login
        - password
    UpdatePasswordDto:
      type: object
      properties:
        oldPassword:
          type: string
          example: test-password
        newPassword:
          type: string
          example: new-password
      required:
        - oldPassword
        - newPassword
    Track:
      type: object
      properties:
        id:
          type: string
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
          format: uuid
        name:
          type: string
          example: Bohemian Rhapsody
        artistId:
          type: string
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
          format: uuid
          nullable: true
        albumId:
          type: string
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
          format: uuid
          nullable: true
        duration:
          type: number
          example: 355
          description: In seconds
      required:
        - id
        - name
        - artistId
        - albumId
        - duration
    CreateTrackDto:
      type: object
      properties:
        name:
          type: string
          example: Bohemian Rhapsody
        artistId:
          type: string
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
          format: uuid
          nullable: true
        albumId:
          type: string
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
          format: uuid
          nullable: true
        duration:
          type: number
          example: 262
          description: In seconds
      required:
        - name
        - artistId
        - albumId
        - duration
    UpdateTrackDto:
      type: object
      properties:
        name:
          type: string
          example: Bohemian Rhapsody
        artistId:
          type: string
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
          format: uuid
          nullable: true
        albumId:
          type: string
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
          format: uuid
          nullable: true
        duration:
          type: number
          example: 262
          description: In seconds
    Artist:
      type: object
      properties:
        id:
          type: string
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
          format: uuid
        name:
          type: string
          example: Freddie Mercury
        grammy:
          type: boolean
          example: false
      required:
        - id
        - name
        - grammy
    CreateArtistDto:
      type: object
      properties:
        name:
          type: string
          example: Freddie Mercury
        grammy:
          type: boolean
          example: false
      required:
        - name
        - grammy
    UpdateArtistDto:
      type: object
      properties:
        name:
          type: string
          example: Freddie Mercury
        grammy:
          type: boolean
          example: false
    Album:
      type: object
      properties:
        id:
          type: string
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
          format: uuid
        name:
          type: string
          example: Innuendo
        year:
          type: number
          example: 1991
        artistId:
          type: string
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
          format: uuid
          nullable: true
      required:
        - id
        - name
        - year
        - artistId
    CreateAlbumDto:
      type: object
      properties:
        name:
          type: string
          example: Innuendo
        year:
          type: number
          example: 1991
        artistId:
          type: string
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
          format: uuid
          nullable: true
      required:
        - name
        - year
        - artistId
    UpdateAlbumDto:
      type: object
      properties:
        name:
          type: string
          example: Innuendo
        year:
          type: number
          example: 1991
        artistId:
          type: string
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
          format: uuid
          nullable: true
    ResponseFavoritesDto:
      type: object
      properties:
        artists:
          example:
            - id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
              name: Freddie Mercury
              grammy: false
          type: array
          items:
            type: string
        albums:
          example:
            - id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
              name: Innuendo
              year: 1991
              artistId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
          type: array
          items:
            type: string
        tracks:
          example:
            - id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
              name: The Show Must Go On
              artistId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
              albumId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
              duration: 262
          type: array
          items:
            type: string
      required:
        - artists
        - albums
        - tracks